refactor shader type enum iteration

+++ AaEngine/source/MaterialResources.cpp
@@ -71,7 +71,7 @@ void MaterialResources::loadMaterials(std::string directory, bool subDirectories
 		MaterialRef& info = knownMaterials[i];
 
 		// post process
-		if (info.pipeline.shaders[(int)ShaderType::Vertex] == "vsQuad")
+		if (info.pipeline.shaders[ShaderType::Vertex] == "vsQuad")
 			continue;
 
 		if (!info.abstract && info.pipeline.fill == D3D12_FILL_MODE_SOLID && !info.techniqueMaterial[int(MaterialTechnique::Wireframe)])
